fixed precision in cython

+++ sklearn/cluster/bicluster/_square_residue.pyx
@@ -18,16 +18,17 @@ ctypedef np.int64_t LONG
 def square_residue(np.ndarray[LONG, ndim=1, mode="c"] rows,
                    np.ndarray[LONG, ndim=1, mode="c"] cols,
                    np.ndarray[DOUBLE, ndim=2, mode="c"] X):
+
     cdef int n_rows = len(rows)
     cdef int n_cols = len(cols)
-    cdef float arr_mean = 0.0
+    cdef double arr_mean = 0.0
 
     cdef np.ndarray[DOUBLE, ndim = 1, mode = "c"] row_mean = None
     cdef np.ndarray[DOUBLE, ndim = 1, mode = "c"] col_mean = None
     row_mean = np.zeros(n_rows, dtype=np.float, order="c")
     col_mean = np.zeros(n_cols, dtype=np.float, order="c")
 
-    cdef float val = 0.0
+    cdef double val = 0.0
 
     for i in range(n_rows):
         for j in range(n_cols):
